refactor: Rename common package to view
