+++ src/main/java/persistence/sql/dml/FindAllQueryBuilder.java
@@ -1,7 +1,7 @@
 package persistence.sql.dml;
 
-import persistence.sql.common.ColumnNames;
-import persistence.sql.common.TableName;
+import persistence.sql.view.ColumnNames;
+import persistence.sql.view.TableName;
 
 public class FindAllQueryBuilder<T> {
     private final Class<T> clazz;
