+++ src/main/java/persistence/sql/view/ColumnValues.java
@@ -1,4 +1,4 @@
-package persistence.sql.common;
+package persistence.sql.view;
 
 import persistence.sql.exception.IllegalFieldAccessException;
 
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static persistence.sql.common.StringConstant.DELIMITER;
+import static persistence.sql.view.StringConstant.DELIMITER;
 
 public class ColumnValues {
     private final List<ColumnValue> values = new ArrayList<>();
